requirements.txt
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.33